docs: update select docs
feat: add support to select for layerRootId
