--- src/misc/select/__docs__/selectStoryHelpers.ts
@@ -1,43 +1,0 @@
-// These fields should not be set via storybook as they require some more complex details.
-export const disabledSelectStoryFields = {
-  children: {
-    table: {
-      disable: true
-    }
-  },
-  isLoading: {
-    table: {
-      disable: true
-    }
-  },
-  hasMore: {
-    table: {
-      disable: true
-    }
-  },
-  onLoadMore: {
-    table: {
-      disable: true
-    }
-  },
-  onSelectClosed: {
-    table: {
-      disable: true
-    }
-  },
-  selectedValues: {
-    table: {
-      disable: true
-    }
-  },
-  onSearchChange: {
-    table: {
-      disable: true
-    }
-  },
-  searchValue: {
-    table: {
-      disable: true
-    }
-  }
-};
